refactor code (cont.)
