+++ NEDAS/dataset/ecmwf/forecast/__init__.py
@@ -1,1 +1,1 @@
-from .dataset import Dataset
+from NEDAS.dataset.ecmwf.forecast.dataset import Dataset
