+++ NEDAS/dataset/ifremer/argo/__init__.py
@@ -1,1 +1,1 @@
-from .dataset import Dataset
+from NEDAS.dataset.ifremer.argo.dataset import Dataset
