+++ NEDAS/dataset/ifremer/argo/dataset.py
@@ -2,8 +2,8 @@ import os
 import glob
 import numpy as np
 from datetime import datetime, timedelta
-from ....utils.conversion import dt1h
-from ...dataset_config import DatasetConfig
+from NEDAS.utils.conversion import dt1h
+from NEDAS.dataset.dataset_config import DatasetConfig
 
 class Dataset(DatasetConfig):
