+++ NEDAS/dataset/osisaf/ice_conc/__init__.py
@@ -1,1 +1,1 @@
-from .dataset import Dataset
+from NEDAS.dataset.osisaf.ice_conc.dataset import Dataset
