+++ NEDAS/dataset/osisaf/ice_drift/__init__.py
@@ -1,1 +1,1 @@
-from .dataset import Dataset
+from NEDAS.dataset.osisaf.ice_drift.dataset import Dataset
