+++ NEDAS/dataset/qg/__init__.py
@@ -1,1 +1,1 @@
-from NEDAS.dataset.qg import Dataset
+from NEDAS.dataset.qg.qg import Dataset
