+++ NEDAS/dataset/vort2d/__init__.py
@@ -1,1 +1,1 @@
-from NEDAS.dataset.vort2d import Dataset
+from NEDAS.dataset.vort2d.vort2d import Dataset
