+++ NEDAS/diag/misc/convert_output/__init__.py
@@ -1,1 +1,1 @@
-from .convert_output import run, get_task_list, get_file_list
+from NEDAS.diag.misc.convert_output.convert_output import run, get_task_list, get_file_list
