+++ NEDAS/models/nextsim/v1/__init__.py
@@ -1,1 +1,1 @@
-from .model import NextsimModel as Model
+from NEDAS.models.nextsim.v1.model import NextsimModel as Model
